requirements.txt
numpy==1.16.4
pandas==0.24.2
opencv-python==4.1.1
scikit-learn=0.21.2
tensorflow==1.13.1
pillow==6.1.0